# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Amrutam Telemedicine API
  version: 1.0.0
  description: |
    Production-grade REST API for an Ayurveda telemedicine platform.

    ## Authentication
    Protected endpoints require a JWT bearer token:
    ```
    Authorization: Bearer <access_token>
    ```
    Obtain tokens via `POST /api/v1/auth/login`.

    ## Idempotency
    `POST /api/v1/appointments` requires an `Idempotency-Key` header (UUID recommended).

    ## Response Envelope
    Success: `{ success, data, meta, requestId, timestamp }`
    Error: `{ success: false, code, message, details, requestId, timestamp }`
  contact:
    name: Amrutam Engineering
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from POST /auth/login
    idempotencyKey:
      type: apiKey
      in: header
      name: Idempotency-Key
      description: Client-generated UUID for idempotent booking
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [success, code, message, requestId, timestamp]
      properties:
        success:
          type: boolean
          const: false
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        details:
          nullable: true
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
...
    PrescriptionItem:
      type: object
      required: [medicineName, strength, dosage, frequency, durationDays]
      properties:
        medicineName:
          type: string
        genericName:
          type: string
        strength:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        durationDays:
          type: integer
          minimum: 1
        instructions:
          type: string
        foodRelation:
          type: string
          enum: [BEFORE_FOOD, AFTER_FOOD, WITH_FOOD, EMPTY_STOMACH]
        notes:
          type: string
    PrescriptionRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PrescriptionItem'
        notes:
          type: string
    InitiatePaymentRequest:
      type: object
      required: [appointmentId]
      properties:
        appointmentId:
          type: string
          format: uuid
        idempotencyKey:
          type: string
    RefundPaymentRequest:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
          minimum: 0.01
        reason:
          type: string
...
  responses:
    Unauthorized:
      description: Invalid or missing bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            code: UNAUTHORIZED
            message: Invalid or missing bearer token
            details: null
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: '2026-07-10T12:00:00.000Z'
    Forbidden:
      description: Insufficient role or resource access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            code: FORBIDDEN
            message: Insufficient role or resource access
            details: null
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: '2026-07-10T12:00:00.000Z'
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            code: VALIDATION_ERROR
            message: Request validation failed
            details: ['doctorId must be a UUID']
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: '2026-07-10T12:00:00.000Z'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            code: NOT_FOUND
            message: Resource not found
            details: null
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: '2026-07-10T12:00:00.000Z'
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            code: INTERNAL_ERROR
            message: An unexpected error occurred
            details: null
            requestId: 550e8400-e29b-41d4-a716-446655440000
            timestamp: '2026-07-10T12:00:00.000Z'
...
paths:
...
  /consultations/{id}/prescription:
    post:
      tags: [Prescriptions]
      summary: Create prescription for consultation
      description: Doctor-only. Creates initial prescription with immutable version history.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrescriptionRequest'
            example:
              items:
                - medicineName: Triphala Churna
                  strength: 500mg
                  dosage: 1 tsp
                  frequency: Twice daily
                  durationDays: 30
                  foodRelation: AFTER_FOOD
      responses:
        '200':
          description: Prescription created
          content:
            application/json:
              example:
                success: true
                data:
                  id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  version: 1
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /prescriptions/{id}:
    get:
      tags: [Prescriptions]
      summary: Get prescription by ID
      description: Returns current version with items. Access restricted by role.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Prescription details
          content:
            application/json:
              example:
                success: true
                data:
                  id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  items: []
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Prescriptions]
      summary: Update prescription
      description: Doctor-only. Creates new immutable version; history preserved.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrescriptionRequest'
      responses:
        '200':
          description: New version created
          content:
            application/json:
              example:
                success: true
                data:
                  version: 2
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /prescriptions/{id}/cancel:
    post:
      tags: [Prescriptions]
      summary: Cancel prescription
      description: Doctor-only. Marks cancelled without deleting version history.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Prescription cancelled
          content:
            application/json:
              example:
                success: true
                data:
                  status: CANCELLED
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Already cancelled
          content:
            application/json:
              example:
                success: false
                code: PRESCRIPTION_CANCELLED
                message: Prescription is already cancelled
        '500':
          $ref: '#/components/responses/InternalError'
  /prescriptions/history/{consultationId}:
    get:
      tags: [Prescriptions]
      summary: Prescription version history
      description: Returns all immutable versions for a consultation.
      parameters:
        - name: consultationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Version history
          content:
            application/json:
              example:
                success: true
                data:
                  versions: []
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /payments/initiate:
    post:
      tags: [Payments]
      summary: Initiate payment for appointment
      description: Creates payment intent via configured provider (MOCK in development). Optional body idempotencyKey.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitiatePaymentRequest'
            example:
              appointmentId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
              idempotencyKey: pay-550e8400-e29b-41d4-a716
      responses:
        '200':
          description: Payment initiated
          content:
            application/json:
              example:
                success: true
                data:
                  paymentId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  status: PENDING
                  amount: 500
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /payments/history:
    get:
      tags: [Payments]
      summary: Get payment history
      description: Paginated payment history for authenticated user.
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Payment history
          content:
            application/json:
              example:
                success: true
                data:
                  items: []
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /payments/{id}:
    get:
      tags: [Payments]
      summary: Get payment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Payment details
          content:
            application/json:
              example:
                success: true
                data:
                  id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  status: CAPTURED
                  amount: 500
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /payments/webhook:
    post:
      tags: [Payments]
      summary: Payment provider webhook
      description: Receives async payment status updates. HMAC-SHA256 verified via `x-webhook-signature` header.
      security: []
      parameters:
        - name: x-webhook-signature
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
            example:
              event: payment.captured
              paymentId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
              amount: 500
      responses:
        '200':
          description: Webhook processed
          content:
            application/json:
              example:
                received: true
        '401':
          description: Invalid webhook signature
          content:
            application/json:
              example:
                success: false
                code: WEBHOOK_VERIFICATION_FAILED
                message: Invalid webhook signature
        '409':
          description: Duplicate webhook
          content:
            application/json:
              example:
                success: false
                code: WEBHOOK_DUPLICATE
                message: Webhook already processed
        '500':
          $ref: '#/components/responses/InternalError'
  /payments/{id}/refund:
    post:
      tags: [Payments]
      summary: Refund a payment
      description: Initiates partial or full refund via payment provider.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundPaymentRequest'
            example:
              amount: 500
              reason: Appointment cancelled
      responses:
        '200':
          description: Refund initiated
          content:
            application/json:
              example:
                success: true
                data:
                  refundId: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                  status: PROCESSING
                requestId: 550e8400-e29b-41d4-a716-446655440000
                timestamp: '2026-07-10T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...